refactor: rename arboretum.common.errors to arboretum.common.exceptions
